Fix bug #4794 (#4808)
* Fix bug #4794

* fix bug #4794

* fix bug #4794

+++ ext-src/swoole_timer.cc
@@ -182,11 +182,12 @@ static void timer_add(INTERNAL_FUNCTION_PARAMETERS, bool persistent) {
     }
 
     // no server || user worker || task process with async mode
-    if (!sw_server() || sw_server()->is_user_worker() || (sw_server()->is_task_worker() && sw_server()->task_enable_coroutine)) {
+    if (!sw_server() || sw_server()->is_user_worker() ||
+        (sw_server()->is_task_worker() && sw_server()->task_enable_coroutine)) {
         php_swoole_check_reactor();
     }
 
-    tnode = swoole_timer_add(ms, persistent, timer_callback, fci);
+    tnode = swoole_timer_add((long) ms, persistent, timer_callback, fci);
     if (UNEXPECTED(!tnode)) {
         php_swoole_fatal_error(E_WARNING, "add timer failed");
         goto _failed;
